server/requirements.txt
agithub==2.1
asyncpg==0.12.0
google-cloud-datastore==1.3.0
requests[security]==2.18.4
tornado==4.5.1
